+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.0 (under development)
 
+* [feature] Add methods hasEdge/getEdge on graph (Alexandre Salomé)
 * [feature] Add methods getAttribute on node and edge (Alexandre Salomé)
 * [feature] Add method get($id) on graph to get a subgraph or a node (Alexandre Salomé)
 * [feature] Disable label escaping (Alexandre Salomé)